feat(openstack.keystone): add dependencies to the tenant, role, and user creation states
Tenants and roles may be created independent of each other and are only
dependent on the Keystone service.  Users must be created after the
tenants and roles of which they are a part, so this state formula now
calculates those dependencies by analyzing the user's `tenant` and
`roles` properties.  The state formula stores the list of required
tenants and roles in dictionaries, which guarantees key uniqueness.